+++ src/changes/revertChange.ts
@@ -19,9 +19,9 @@ import { Transaction } from 'prosemirror-state'
 import { liftTarget, Mapping, ReplaceAroundStep } from 'prosemirror-transform'
 
 import { ChangeSet } from '../ChangeSet'
-import { getBlockInlineTrackedData } from '../compute/nodeHelpers'
 import { CHANGE_OPERATION, CHANGE_STATUS, IncompleteChange, NodeChange, TrackedChange } from '../types/change'
 import { excludeFromTracked } from '../utils/tracking'
+import { getBlockInlineTrackedData } from '../attributes'
 
 /**
  *  move split-ed content back to the original node. and will update original node dataTracked in these cases:
