+++ src/steps-trackers/lib/deleteAndMergeSplitNodes.ts
