+++ src/steps-trackers/lib/deleteText.ts
@@ -16,8 +16,8 @@
 import { Fragment, Node as PMNode, Schema } from 'prosemirror-model'
 import type { Transaction } from 'prosemirror-state'
 
-import { addTrackIdIfDoesntExist, getMergeableMarkTrackedAttrs } from '../compute/nodeHelpers'
 import { NewDeleteAttrs } from '../attributes/types'
+import { addTrackIdIfDoesntExist, getMergeableMarkTrackedAttrs } from '../attributes'
 
 /**
  * Deletes inserted text directly, otherwise wraps it with tracked_delete mark
