+++ src/steps-trackers/lib/mergeNode.ts
